Introduce tpolecat sbt plugin, resolve compiler warnings (#199)
* Introduce tpolecat sbt plugin, resolve compiler warnings

* Build doc in travis, fix scaladocs

+++ www/app/controllers/ApplicationController.scala
@@ -2,7 +2,8 @@ package controllers
 
 import io.flow.dependency.www.lib.DependencyClientProvider
 import io.flow.play.controllers.FlowControllerComponents
-import io.flow.play.util.{Config, PaginatedCollection, Pagination}
+import io.flow.play.util.{PaginatedCollection, Pagination}
+import io.flow.util.Config
 import play.api.mvc._
 
 import scala.concurrent.ExecutionContext
@@ -24,8 +25,8 @@ class ApplicationController @javax.inject.Inject()(
     for {
       recommendations <- dependencyClient(request).recommendations.get(
         organization = organization,
-        limit = Pagination.DefaultLimit + 1,
-        offset = page * Pagination.DefaultLimit
+        limit = Pagination.DefaultLimit.toLong + 1L,
+        offset = page * Pagination.DefaultLimit.toLong
       )
     } yield {
       Ok(
